Meta: overhaul application structure
- use create_app() instead of global app object
- introduce 'main' blueprint to get rid of default routes
- move functions around to accomodate for the changes
