+++ app/main/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+bp = Blueprint("main", __name__)
+
+from app.main import routes
