+++ app/static/js/helpers/datatables.js
