+++ app/static/js/helpers/multi-select.js
