+++ app/templates/about.html
@@ -24,7 +24,7 @@
         You can contact me via <a href="https://github.com/tarenethil">github</a> or <a href="https://old.reddit.com/u/TarEnethil">reddit</a>.
     </p>
 
-    <img id="logo" alt="Archivar logo" src="{{ url_for('static_files', filename='logo.png') }}">
+    <img id="logo" alt="Archivar logo" src="{{ url_for('static', filename='logo.png') }}">
 
     <!-- <h2>Special thanks</h2> -->
