+++ app/templates/footer.html
@@ -1,6 +1,6 @@
 <p id="footer">
     version {{ get_archivar_version() }}
-    &bullet; <a href="{{ url_for('about') }}">About Archivar</a>
-    &bullet; <a href="{{ url_for('changelog') }} ">Changelog</a>
-    &bullet; <a href="{{ url_for('statistics') }}">Statistics</a>
+    &bullet; <a href="{{ url_for('main.about') }}">About Archivar</a>
+    &bullet; <a href="{{ url_for('main.changelog') }} ">Changelog</a>
+    &bullet; <a href="{{ url_for('main.statistics') }}">Statistics</a>
 </p>
