+++ app/templates/index.html
@@ -4,7 +4,7 @@
 {% if current_user.has_admin_role() %}
     <ul class="nav nav-tabs">
         <li>
-            <a href="{{ url_for('settings') }}">
+            <a href="{{ url_for('main.settings') }}">
                 {{ icon("pencil") }}
                 Edit Text
             </a>
