+++ app/templates/map/index.html
@@ -3,7 +3,7 @@
 
 {% block styles %}
 {{ super() }}
-<link rel="stylesheet" type="text/css" href="{{ url_for('static_files', filename='map.css') }}">
+<link rel="stylesheet" type="text/css" href="{{ url_for('static', filename='map.css') }}">
 {{ include_css(['leaflet', 'bootstrap-select', 'markdown-editor']) }}
 {% endblock %}
