+++ app/templates/wiki/base.html
@@ -4,7 +4,7 @@
 
 {% block styles %}
 {{ super() }}
-<link rel="stylesheet" type="text/css" href="{{ url_for('static_files', filename='wiki.css') }}">
+<link rel="stylesheet" type="text/css" href="{{ url_for('static', filename='wiki.css') }}">
 {% endblock %}
 
 {% block scripts %}
